175: Rename data service and fix typos
